pyproject.toml
[project]
name = "pyconjp-image-search"
version = "0.1.0"
description = "PyCon JP image search: download, manage, and browse event photos"
requires-python = ">=3.13"

dependencies = [
    "duckdb>=1.2",
    "gradio>=5.0",
    "httpx>=0.28",
    "numpy>=1.26",
    "pillow>=11.0",
    "protobuf>=5.0",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
    "rich>=13.0",
    "sentencepiece>=0.2",
    "torch>=2.0",
    "transformers>=4.40",
]

[project.scripts]
pyconjp-manage = "pyconjp_image_search.manager:main"
pyconjp-embed = "pyconjp_image_search.embedding:main"
pyconjp-search = "pyconjp_image_search.search:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "pyright>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyconjp_image_search"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py313"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
